pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-data-router-agent"
version = "0.1.0"
description = "AI Agent system based on Router-Retriever-Generator pattern for MongoDB time-series data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Developer"}
]

dependencies = [
    "chromadb>=0.5.0",
    "pymongo>=4.6.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "redis>=5.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "python-dateutil>=2.8.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langchain-openai>=0.0.5",
    "httpx>=0.26.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.92.0",
    "mongomock>=4.1.0",
    "fakeredis>=2.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/manual"]
cache_dir = ".cache/pytest"
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hypothesis]
max_examples = 100
